+++ tests/metrics/qualification/cli/vehicle_role/test_get_default.py
@@ -1,0 +1,7 @@
+"""."""
+
+from omega_prime.metrics.qualification.cli.vehicle_role import VehicleRole
+
+
+def test_get_default() -> None:
+    assert VehicleRole.get_default() == tuple()