+++ tests/metrics/qualification/cli/vehicle_type/test_get_default.py
@@ -1,0 +1,7 @@
+"""."""
+
+from omega_prime.metrics.qualification.cli.vehicle_type import VehicleType
+
+
+def test_get_default() -> None:
+    assert VehicleType.get_default() == tuple()